+++ Octokit.Reactive/Clients/IObservableIssuesEventsClient.cs
@@ -20,7 +20,7 @@ namespace Octokit.Reactive
         /// <param name="owner">The owner of the repository</param>
         /// <param name="name">The name of the repository</param>
         /// <param name="issueNumber">The issue number</param>
-        IObservable<IssueEvent> GetAllForIssue(string owner, string name, int issueNumber);
+        IObservable<IssueEvent> GetAllForIssue(string owner, string name, long issueNumber);
 
         /// <summary>
         /// Gets all events for the issue.
@@ -30,7 +30,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="issueNumber">The issue number</param>
-        IObservable<IssueEvent> GetAllForIssue(long repositoryId, int issueNumber);
+        IObservable<IssueEvent> GetAllForIssue(long repositoryId, long issueNumber);
 
         /// <summary>
         /// Gets all events for the issue.
@@ -42,7 +42,7 @@ namespace Octokit.Reactive
         /// <param name="name">The name of the repository</param>
         /// <param name="issueNumber">The issue number</param>
         /// <param name="options">Options for changing the API response</param>
-        IObservable<IssueEvent> GetAllForIssue(string owner, string name, int issueNumber, ApiOptions options);
+        IObservable<IssueEvent> GetAllForIssue(string owner, string name, long issueNumber, ApiOptions options);
 
         /// <summary>
         /// Gets all events for the issue.
@@ -53,7 +53,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="issueNumber">The issue number</param>
         /// <param name="options">Options for changing the API response</param>
-        IObservable<IssueEvent> GetAllForIssue(long repositoryId, int issueNumber, ApiOptions options);
+        IObservable<IssueEvent> GetAllForIssue(long repositoryId, long issueNumber, ApiOptions options);
 
         /// <summary>
         /// Gets all events for the repository.
